Extract Modal to Ballot
